refactor: split reusable components
